+++ app/renderer/src/main/src/store/editorFontSize.ts
@@ -15,23 +15,31 @@ interface EditorFontSizeStore {
 
 export const fontSizeOptions = [12, 13, 14, 15, 16, 17, 18, 19, 20]
 
+let fontSizeInitPromise: Promise<void> | null = null
+
 export const useEditorFontSize = create<EditorFontSizeStore>((set, get) => ({
   fontSize: 12,
   setFontSize: (fontSize) => {
+    if (get().fontSize === fontSize) return
     set({ fontSize })
     setRemoteValue(HTTP_PACKET_EDITOR_FONT_SIZE, JSON.stringify({ fontSize }))
   },
   initFontSize: async () => {
-    try {
-      const data = await getRemoteValue(HTTP_PACKET_EDITOR_FONT_SIZE)
-      if (data) {
-        const obj = JSONParseLog(data, { page: 'editorFontSize', fun: 'initFontSize' })
-        if (obj?.fontSize && obj.fontSize >= 12 && obj.fontSize <= 20) {
-          set({ fontSize: obj.fontSize })
+    // 全局只初始化一次，避免多个编辑器同时 init 互相触发重渲染
+    if (fontSizeInitPromise) return fontSizeInitPromise
+    fontSizeInitPromise = (async () => {
+      try {
+        const data = await getRemoteValue(HTTP_PACKET_EDITOR_FONT_SIZE)
+        if (data) {
+          const obj = JSONParseLog(data, { page: 'editorFontSize', fun: 'initFontSize' })
+          if (obj?.fontSize && obj.fontSize >= 12 && obj.fontSize <= 20 && get().fontSize !== obj.fontSize) {
+            set({ fontSize: obj.fontSize })
+          }
         }
+      } catch (error) {
+        yakitFailed(error + '')
       }
-    } catch (error) {
-      yakitFailed(error + '')
-    }
+    })()
+    return fontSizeInitPromise
   },
 }))
